Time: 18 ms (9.95%), Space: 23 MB (58.74%) - LeetHub - Time: O(N * log(max(piles))), Space: O(1)

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -10,7 +10,7 @@ int findmax(vector<int>& piles){
 long long K(vector<int>& piles , int k){
     long long time = 0 ;
         for(int i = 0 ; i <piles.size() ; i++){
-          time += ceil((double)piles[i]/k);
+          time += ceil((double)piles[i]/(double)k);
         }
         return time; 
 }
